Major refactor in code structure, renamed Topic to Frame and separated the code accordingly
