+++ src/assets/scss/components/_frame.scss
@@ -1,4 +1,4 @@
-.topic {
+.frame {
     display: grid;
     line-height: 1.5;
     height: min-content;
