+++ src/assets/scss/pages/_post.scss
@@ -4,7 +4,7 @@
     justify-content: center;
     margin-bottom: 2rem;
 
-    .topic {
+    .frame {
         &.original {
             width: 50%;
