+++ src/assets/scss/pages/_topics.scss
@@ -1,5 +1,5 @@
 .tree {
-    .topic {
+    .frame {
         .body {
             display: grid;
             gap: 1rem;
