+++ src/components/NoResponse.jsx
@@ -1,0 +1,5 @@
+export default () => (
+  <div className="no-response">
+    ainda não há repostas, que tal contribuir?
+  </div>
+)
