+++ src/components/Topic.jsx
@@ -1,129 +1,89 @@
 import React from "react"
-import { isEmptyObject } from "@tiptap/react"
-import { PiDotsThreeVerticalBold } from "react-icons/pi"
+import { Link, useNavigate } from "react-router-dom"
+import {
+  PiBookmarkSimple,
+  PiBookmarkSimpleFill,
+  PiArrowBendUpLeft,
+  PiCaretDoubleUp
+} from "react-icons/pi"
 
-import useScreenSize from "@/hooks/useScreenSize"
-import VotingButtons from "@/components/VotingButtons"
+import Frame from "@/components/Frame"
+import PostSummary from "@/components/PostSummary"
+import NoResponse from "@/components/NoResponse"
+import { submitVote } from "@/components/VotingButtons"
+import { toPercentageStr } from "@/assets/util"
 
 
 export default function Topic({
   id,
+  author,
   title,
-  setTitle,
-  metrics = [],
-  headerConfig = {},
-  saveInLocalStorage = false,
-  readOnly = true,
-  hideVoteButtons = false,
-  showDefinitiveVoteButton = false,
-  initialVoteState = { vote: false, relevance: "" },
-  alongsideCritique = false,
-  isCritique = false,
-  justify = false,
-  error = false,
-  setHeight = () => { },
+  promotions,
+  upvotes,
+  downvotes,
+  config,
   children,
-  ...remainingProps
+  childrenStats,
+  userInteractions,
+  userVote
 }) {
-  const [headerStatus, setHeaderStatus] = React.useState(
-    Object.fromEntries(
-      Object.entries(headerConfig)
-        .map(([k, v]) => [k, v.initialValue])
-        .filter((value) => value !== undefined)
-    )
-  )
-  const topicRef = React.useRef()
-  const titleRef = React.useRef()
-  const isDesktop = useScreenSize()
+  const [relevanceVote, setRelevanceVote] = React.useState(userInteractions?.filter(v => v === "up" || v === "down")[0])
+  const [definitiveVote, setDefinitiveVote] = React.useState(userVote)
+  const navigate = useNavigate()
 
-  const toggle = (str) => { setHeaderStatus({ ...headerStatus, [str]: !headerStatus[str] }) }
-
-
-  React.useEffect(() => {
-    if (titleRef.current) {
-      titleRef.current.style.height = '32px';
-      titleRef.current.style.height = `${titleRef.current.scrollHeight + 2}px`;
+  const headerConfig = {
+    "answer": {
+      description: "responder tópico",
+      icons: PiArrowBendUpLeft,
+      onClick: () => navigate("/write", { state: { id, title, config } })
+    },
+    "promote": {
+      description: "promover tópico",
+      icons: PiCaretDoubleUp,
+      onClick: () => submitVote(navigate, id, "promote",)
+    },
+    "bookmark": {
+      description: ["salvar tópico", "remover tópico dos salvos"],
+      icons: [PiBookmarkSimple, PiBookmarkSimpleFill],
+      initialValue: userInteractions?.includes("bookmark") || false,
+      onClick: () => submitVote(navigate, id, "bookmark")
     }
-  }, [title]);
+  }
 
+  const getMetrics = () => {
+    const allVotes = upvotes + downvotes
+    const interactions = childrenStats?.upvotes + childrenStats?.downvotes
 
-  React.useEffect(() => {
-    setHeight(topicRef.current.clientHeight || 0)
-  }, [])
-
+    return [
+      `iniciado por ${author}`,
+      `promovido por ${promotions} usuário${promotions === 1 ? "" : "s"}`,
+      allVotes ? `${toPercentageStr(upvotes / allVotes)} dos ${allVotes} votantes achou relevante` : "0 votos",
+      `${childrenStats?.count} post${childrenStats?.count === 1 ? "" : "s"}`,
+      `${interactions} interaç${interactions === 1 ? "ão" : "ões"}`
+    ]
+  }
 
   return (
-    <div className={`topic${alongsideCritique ? " original" : ""}${isCritique ? " critique" : ""}`} ref={topicRef} {...remainingProps}>
-      <div className="header">
-        <div className={`top bracket${error ? " error" : ""}`} />
-        {!hideVoteButtons &&
-          <VotingButtons id={id} showDefinitiveVoteButton={showDefinitiveVoteButton} initialState={initialVoteState} />
-        }
-        <h1 className="title">
-          {readOnly ?
-            title
-            :
-            <textarea
-              placeholder="Qual é o título?"
-              value={title}
-              onChange={e => setTitle(e.target.value)}
-              onKeyDown={e => { if (e.key === "Enter") { editor.chain().focus().run(); e.preventDefault() } }}
-              onBlur={_ => saveInLocalStorage && localStorage.setItem("postTitle", title)}
-              ref={titleRef}
-            />
-          }
-        </h1>
-        {!isEmptyObject(headerConfig) &&
-          <div className="buttons">
-            {isDesktop ?
-              <>
-                {Object.entries(headerConfig).map((([buttonName, buttonConfig]) => {
-                  switch (headerStatus[buttonName]) {
-                    case false:
-                      return buttonConfig.icons[0]({
-                        className: "icons",
-                        title: buttonConfig.description[0],
-                        onClick: () => { buttonConfig.onClick(); toggle(buttonName) }
-                      })
-                    case true:
-                      return buttonConfig.icons[1]({
-                        className: "icons",
-                        title: buttonConfig.description[1],
-                        onClick: () => { buttonConfig.onClick(); toggle(buttonName) }
-                      })
-                    case undefined:
-                      return buttonConfig.icons({
-                        className: "icons",
-                        title: buttonConfig.description,
-                        onClick: () => buttonConfig.onClick()
-                      })
-                  }
-                }))}
-              </>
-              :
-              <PiDotsThreeVerticalBold className="icons" />
-            }
-          </div>
-        }
-      </div>
-      <div className="container">
-        <div className={`bracket${error ? " error" : ""}`} />
-        <div className={`body${justify ? " justify" : ""}`}>
-
-          {children.constructor === Array ?
-            children
-            :
-            React.cloneElement(children, { ...headerStatus, readOnly, saveInLocalStorage, alongsideCritique })
-          }
-          {metrics &&
-            <ul className="metrics">
-              {metrics.map(metric => (
-                <li>{metric}</li>
-              ))}
-            </ul>
-          }
-        </div>
-      </div>
-    </div>
+    <Frame
+      id={id}
+      title={<Link to={`/topics/${id}`}>{String(title)}</Link>}
+      initialVoteState={{ relevance: relevanceVote, vote: definitiveVote }}
+      headerConfig={headerConfig}
+      metrics={getMetrics()}
+    >
+      {children?.length > 0 ?
+        children.map(child => (
+          <PostSummary
+            parent_id={id}
+            id={child.id}
+            shortAnswer={child.title}
+            percentage={child.upvotes / childrenStats?.upvotes}
+            chosen={userVote === child.id}
+          />
+        ))
+        :
+        <NoResponse />
+      }
+    </Frame>
   )
 }
